No need to run darwin tests for mas builds, it's the same codepath
